+++ NEWS.md
@@ -1,4 +1,9 @@
 # agriutilities 1.0.0.9000
 
+## Minor changes/fixes
+
+  * Add S3 method to plot an object of class smaAgri `plot.smaAgri()`.
+  * Fix argument not working in `covcor_heat()`.
+
 # agriutilities 1.0.0